android wallet native 1.0.0

+++ particle-wallet/android/src/main/java/com/particlewallet/model/InitData.java
@@ -10,7 +10,7 @@ public class InitData {
     public String chainName;
 
     @SerializedName("chain_id")
-    public int chainId;
+    public long chainId;
 
     @SerializedName("chain_id_name")
     public String chainIdName;
